pyproject.toml
[project]
name = "gpuport-data-collectors"
version = "0.1.0"
description = "GPUPort Data Collectors"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "GPUPort Team"}
]
dependencies = [
    "aiohttp>=3.13.2",
    "click>=8.3.0",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp-proto-http>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "pydantic>=2.0.0",
    "pydantic-yaml>=1.6.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
gpuport-collectors = "gpuport_collectors.cli:cli"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
    "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gpuport_collectors"]

[tool.hatch.build.targets.wheel.force-include]
"src/gpuport_collectors/py.typed" = "gpuport_collectors/py.typed"

# Dependency groups for uv
[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
    "types-pyyaml>=6.0.0",
    "types-click>=7.1.8",
]

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
# Enable recommended rules plus additional quality checks
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
]

ignore = [
    "E501",   # Line too long (handled by formatter)
    "COM812", # Missing trailing comma (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",  # Unused function arguments in tests (fixtures)
    "S101", # Use of assert
]
"examples/**/*.py" = [
    "T20",  # Allow print statements in examples
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# Mypy configuration
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_any_generics = true
disallow_subclassing_any = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true
exclude = ["examples/"]

# Click decorators are inherently untyped - allow untyped decorators in CLI modules
[[tool.mypy.overrides]]
module = "gpuport_collectors.cli"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gpuport_collectors.observability"
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = "aiohttp"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "click"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydantic_yaml"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=gpuport_collectors",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
